Close the libc and type-checker seams that broke the Linux build
Nine independent stops, each of which halted the whole graph:

- OpenGrokAuth: `#if os(macOS) || os(Linux) import Darwin`. The `Darwin.bind`
  qualification existed only to disambiguate from a local `bind(port:)`, so
  renaming that to `bindListener(port:)` lets plain libc resolve on both
  platforms and the conditional import disappears. `listen`'s status was also
  discarded — a listener that failed to listen was returned as bound.
- OpenGrokTTY/OpenGrokPTY: `ioctl`'s request argument is `UInt`, but glibc
  imports TIOCGWINSZ/TIOCSWINSZ as `Int32`.
- OpenGrokTTY: glibc exports `stderr` as a *mutable* global, which Swift 6
  strict concurrency rejects. `fflush(nil)` off Darwin.
- OpenGrokGitStatus: glibc spells the constant `__RLIMIT_NPROC` and makes
  RLIMIT_NPROC a macro alias that ClangImporter does not surface.
- OpenGrokConfig/FsAtomic: `strdup` is optional-returning on Linux; the whole
  strdup/free dance was unnecessary next to `withCString`.
- OpenGrokConfig/Ed25519Portable and OpenGrokWorkspace/BashSplit: two
  expressions exceeded the Linux type checker's budget. Both rewrites are
  value-identical; `dangerousCommands` is set-equal and macOS tests cover it.
- OpenGrokSampler: bare `import os` for a single `.debug` line.

+++ Sources/OpenGrokPTY/OpenGrokPTY.swift
@@ -500,7 +500,7 @@ public final class PosixPTYProcess: PTYProcess, @unchecked Sendable {
         ws.ws_col = UInt16(clamping: size.width)
         ws.ws_xpixel = 0
         ws.ws_ypixel = 0
-        let rc = ioctl(masterFD, TIOCSWINSZ, &ws)
+        let rc = ioctl(masterFD, UInt(TIOCSWINSZ), &ws)
         if rc != 0 {
             throw PTYError.ioFailed("TIOCSWINSZ failed: \(String(cString: strerror(errno)))")
         }
